refactor: Andrew [#84] move getAggregateVisualizationLayout into the visualization class
